[refactor]#94:createBoard 지원대상 추가 및 api resolve/reject 방법으로 변경
